drishti-mother-child/src/test/java/org/ei/drishti/service/formSubmissionHandler/FormSubmissionRouterTest.java
package org.ei.drishti.service.formSubmissionHandler;

import org.ei.drishti.dto.form.FormSubmission;
import org.junit.Before;
import org.junit.Test;
import org.mockito.Mock;

import static org.mockito.Mockito.verify;
import static org.mockito.MockitoAnnotations.initMocks;
import static org.motechproject.util.DateUtil.now;

public class FormSubmissionRouterTest {
    @Mock
    private ECRegistrationHandler ecRegistrationHandler;
    @Mock
    private FPComplicationsHandler fpComplicationsHandler;

    private FormSubmissionRouter router;

    @Before
    public void setUp() throws Exception {
        initMocks(this);
        router = new FormSubmissionRouter(ecRegistrationHandler, fpComplicationsHandler);
    }

    @Test
    public void shouldDelegateFormSubmissionHandlingToHandlerBasedOnFormName() throws Exception {
        FormSubmission formSubmission = new FormSubmission("anm id 1", "instance id 1", "entity id 1", "ec_registration", null, now().getMillis());

        router.route(formSubmission);

        verify(ecRegistrationHandler).handle(formSubmission);

        formSubmission = new FormSubmission("anm id 1", "instance id 2", "entity id 2", "fp_complications", null, now().getMillis());

        router.route(formSubmission);

        verify(fpComplicationsHandler).handle(formSubmission);
    }
}
